execution/python/requirements.txt
numpy<2
fasttext-wheel==0.9.2
